+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - A `BlockstackConfig` object to make it easier to use across several activities. Thanks
 to @friedger.
 - `encryptContent`, `decryptContent` methods
+- `BlockstackSession.lookupProfile` method
 ### Changed
 - Fixed a bug where loadUserData would throw an exception if the user is not logged in.
 Thanks to @friedger.